// src/types/profilCdcFti.ts
export interface ProfilCdcFtiItem {
  profil_cdc_fti_id: number;
  profil_cdc_fti_firstname: string;
  profil_cdc_fti_lastname: string;
  profil_cdc_fti_img: string;
  profil_cdc_fti_vision_name: string;
  profil_cdc_fti_vision_img: string;
  profil_cdc_fti_mission_name: string;
  profil_cdc_fti_mission_img: string;
  profil_cdc_fti_goal_name: string;
  profil_cdc_fti_goal_img: string;
  profil_cdc_fti_benefit_name: string;
  profil_cdc_fti_benefit_img: string;
  profil_cdc_fti_email: string;
  profil_cdc_fti_instagram: string;
  profil_cdc_fti_youtube: string;
  profil_cdc_fti_facebook: string;
  profil_cdc_fti_whatsapp: string;
  profil_cdc_fti_created_at: string;
  profil_cdc_fti_updated_at: string;
}